Initial monster update/repolib 3.0 fix
Also disabled MoreUpdgrades and VanillaUpgrades compat patches. MU got an overhaul to use REPOLib and thus needs to be revisited, and VU isn't updated to 3.0 yet.
Currently the only issue that has shown up is that the prices are WAY too expensive on bundles. Current theory is price alterations from here and MU are stacking.